# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/users/":{"put":{"tags":["user-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterUserResponse"}}}}}},"post":{"tags":["user-controller"],"operationId":"requestAutoUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestAutoUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestAutoUserResponse"}}}}}}},"/events/{id}":{"put":{"tags":["event-controller"],"operationId":"updateEvent","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericEventResponse"}}}}}}},"/media-items/":{"post":{"tags":["media-item-controller"],"operationId":"uploadNewMediaItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequest"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/events/":{"post":{"tags":["event-controller"],"operationId":"createEvent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericEventResponse"}}}}}}},"/media-items/{hash}":{"get":{"tags":["media-item-controller"],"operationId":"getImageByHash","parameters":[{"name":"hash","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}}},"components":{"schemas":{"RegisterUserRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":10,"type":"string"}}},"RegisterUserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"email":{"type":"string"},"registrationState":{"type":"string"}}},"CreateEventRequest":{"type":"object","properties":{"displayName":{"type":"string"},"description":{"type":"string"},"startDateTime":{"type":"string","format":"date-time"},"duration":{"type":"integer","format":"int32"}}},"GenericEventResponse":{"type":"object","properties":{"displayName":{"type":"string"},"description":{"type":"string"},"startDateTime":{"type":"string","format":"date-time"},"duration":{"type":"integer","format":"int32"},"state":{"type":"string","enum":["PLANNED","ONGOING","PAST"]}}},"RequestAutoUserRequest":{"type":"object","properties":{"name":{"maxLength":100,"minLength":2,"type":"string"}}},"RequestAutoUserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"registrationState":{"type":"string"}}},"CreateRequest":{"type":"object","properties":{"dateTime":{"type":"string","format":"date-time"},"imgDataBase64":{"type":"string"},"eventId":{"type":"integer","format":"int32"},"creatorId":{"type":"integer","format":"int32"}}}}}}